+++ yarn-project/simulator/src/public/avm/avm_simulator.test.ts
@@ -10,6 +10,8 @@ import {
 } from '@aztec/foundation/crypto';
 import { Fq, Fr, Point } from '@aztec/foundation/fields';
 import type { Fieldable } from '@aztec/foundation/serialize';
+import { AvmGadgetsTestContract } from '@aztec/noir-test-contracts.js/AvmGadgetsTest';
+import { AvmTestContract } from '@aztec/noir-test-contracts.js/AvmTest';
 import { CounterContract } from '@aztec/noir-test-contracts.js/Counter';
 import { type FunctionArtifact, FunctionSelector } from '@aztec/stdlib/abi';
 import { AztecAddress } from '@aztec/stdlib/aztec-address';
@@ -39,10 +41,8 @@ import type { AvmContext } from './avm_context.js';
 import type { AvmExecutionEnvironment } from './avm_execution_environment.js';
 import { type MemoryValue, TypeTag, type Uint8, type Uint64 } from './avm_memory_types.js';
 import { AvmSimulator } from './avm_simulator.js';
+import { AvmRevertReason } from './errors.js';
 import {
-  getAvmGadgetsTestContractBytecode,
-  getAvmTestContractArtifact,
-  getAvmTestContractBytecode,
   getContractFunctionArtifact,
   initContext,
   initExecutionEnvironment,
@@ -52,7 +52,6 @@ import {
   randomMemoryBytes,
   randomMemoryFields,
   randomMemoryUint64s,
-  resolveAvmTestContractAssertionMessage,
   resolveContractAssertionMessage,
 } from './fixtures/index.js';
 import {
@@ -472,7 +471,7 @@ describe('AVM simulator: transpiled Noir contracts', () => {
       const calldata = input.map(e => e.toFr());
 
       const context = initContext({ env: initExecutionEnvironment({ calldata }) });
-      const bytecode = getAvmGadgetsTestContractBytecode(name);
+      const bytecode = getAvmGadgetsTestContractArtifact(name).bytecode;
       const results = await new AvmSimulator(context).executeBytecode(bytecode);
 
       expect(results.reverted).toBe(false);
@@ -1307,4 +1306,35 @@ async function pedersenFromMemoryFields(fields: Fieldable[]): Promise<Fr[]> {
 
 async function indexedPedersenFromMemoryFields(fields: Fieldable[]): Promise<Fr[]> {
   return [await pedersenHash(fields, /*index=*/ 20)];
+}
+
+function resolveAvmTestContractAssertionMessage(
+  functionName: string,
+  revertReason: AvmRevertReason,
+  output: Fr[],
+): string | undefined {
+  return resolveContractAssertionMessage(functionName, revertReason, output, AvmTestContract.artifactForPublic);
+}
+
+function getAvmTestContractBytecode(functionName: string): Buffer {
+  const artifact = getAvmTestContractArtifact(functionName);
+  return artifact.bytecode;
+}
+
+function getAvmGadgetsTestContractArtifact(functionName: string): FunctionArtifact {
+  const artifact = AvmGadgetsTestContract.artifactForPublic.functions.find(f => f.name === functionName)!;
+  assert(
+    !!artifact?.bytecode,
+    `No bytecode found for function ${functionName}. Try re-running bootstrap.sh on the repository root.`,
+  );
+  return artifact;
+}
+
+function getAvmTestContractArtifact(functionName: string): FunctionArtifact {
+  const artifact = getContractFunctionArtifact(functionName, AvmTestContract.artifactForPublic) as FunctionArtifact;
+  assert(
+    !!artifact?.bytecode,
+    `No bytecode found for function ${functionName}. Try re-running bootstrap.sh on the repository root.`,
+  );
+  return artifact;
 }
